[Frontend] cleanup fusion code
